+++ src/app/home/management/components/forms/extra-income-form.tsx
@@ -10,6 +10,7 @@ import { useSession } from "@/shared/libs/better-auth/auth-client";
 import { Income } from "@/shared/interfaces/income";
 import { Loader2, Trash } from "lucide-react";
 import { formatDateBR } from "@/shared/helpers/date";
+import { confirmToast, errorToast, loadingAlert, successToast } from "@/shared/libs/react-hot-toast/react-hot-toast";
 
 export default function ExtraIncomeForm() {
   const router = useRouter();
@@ -54,6 +55,7 @@ export default function ExtraIncomeForm() {
       });
 
       reset();
+      successToast("Registro extra adicionado com sucesso!");
       router.push("/home/management");
       router.refresh();
     } catch (err) {
@@ -72,7 +74,7 @@ export default function ExtraIncomeForm() {
       setIncomes(response.data || []);
       setIsLoadingIncomes(false);
     } catch (error) {
-      console.error("Erro ao buscar receitas:", error);
+      errorToast("Erro ao recuperar registros");
     } finally {
       setIsLoadingIncomes(false);
     }
@@ -80,12 +82,16 @@ export default function ExtraIncomeForm() {
 
   const deleteIncome = async (id: string) => {
     if (!session?.user?.id) return;
-    try {
-      await api.delete("/api/income", { data: { id } });
-      fetchIncomes();
-    } catch (error) {
-      console.error("Erro ao deletar receita:", error);
-    }
+  
+    confirmToast("Deseja realmente excluir?", async () => {
+      const promise = api.delete ("/api/income", { data: { id } });
+      loadingAlert("Excluindo registro...", "Registro excluído com sucesso!", "Erro ao excluir registro", promise);
+      promise.then((response) => {
+        if (response.status === 200) {
+          fetchIncomes();
+        }
+      })
+    });
   };
 
   return (
